# internal/trainer/ports/http/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.2 DO NOT EDIT.
package http

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Date defines model for Date.
type Date struct {
	Date         openapi_types.Date `json:"date"`
	HasFreeHours bool               `json:"hasFreeHours"`
	Hours        []Hour             `json:"hours"`
}

// ErrorDetails defines model for ErrorDetails.
type ErrorDetails struct {
	EntityId   *string `json:"entity_id,omitempty"`
	EntityType *string `json:"entity_type,omitempty"`
	ErrorSlug  string  `json:"error_slug"`
	Message    string  `json:"message"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Details Detailed error information
	Details *[]ErrorDetails `json:"details,omitempty"`

	// Message Human-readable error message
	Message string `json:"message"`

	// Slug Error slug
	Slug string `json:"slug"`
}

// Hour defines model for Hour.
type Hour struct {
	Available            bool      `json:"available"`
	HasTrainingScheduled bool      `json:"hasTrainingScheduled"`
	Hour                 time.Time `json:"hour"`
}

// HourUpdate defines model for HourUpdate.
type HourUpdate struct {
	Hours []time.Time `json:"hours"`
}

// bearerAuthContextKey is the context key for bearerAuth security scheme
type bearerAuthContextKey string

// GetTrainerAvailableHoursParams defines parameters for GetTrainerAvailableHours.
type GetTrainerAvailableHoursParams struct {
	DateFrom time.Time `form:"dateFrom" json:"dateFrom"`
	DateTo   time.Time `form:"dateTo" json:"dateTo"`
}

// MakeHourAvailableJSONRequestBody defines body for MakeHourAvailable for application/json ContentType.
type MakeHourAvailableJSONRequestBody = HourUpdate

// MakeHourUnavailableJSONRequestBody defines body for MakeHourUnavailable for application/json ContentType.
type MakeHourUnavailableJSONRequestBody = HourUpdate

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /trainer/calendar)
	GetTrainerAvailableHours(ctx echo.Context, params GetTrainerAvailableHoursParams) error

	// (PUT /trainer/calendar/make-hour-available)
	MakeHourAvailable(ctx echo.Context) error

	// (PUT /trainer/calendar/make-hour-unavailable)
	MakeHourUnavailable(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetTrainerAvailableHours converts echo context to params.
func (w *ServerInterfaceWrapper) GetTrainerAvailableHours(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTrainerAvailableHoursParams
	// ------------- Required query parameter "dateFrom" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "dateFrom", ctx.QueryParams(), &params.DateFrom, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter dateFrom: %s", err))
	}

	// ------------- Required query parameter "dateTo" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, true, "dateTo", ctx.QueryParams(), &params.DateTo, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter dateTo: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetTrainerAvailableHours(ctx, params)
	return err
}

// MakeHourAvailable converts echo context to params.
func (w *ServerInterfaceWrapper) MakeHourAvailable(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.MakeHourAvailable(ctx)
	return err
}

// MakeHourUnavailable converts echo context to params.
func (w *ServerInterfaceWrapper) MakeHourUnavailable(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.MakeHourUnavailable(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(options.BaseURL+"/trainer/calendar", wrapper.GetTrainerAvailableHours, options.OperationMiddlewares["getTrainerAvailableHours"]...)
	router.PUT(options.BaseURL+"/trainer/calendar/make-hour-available", wrapper.MakeHourAvailable, options.OperationMiddlewares["makeHourAvailable"]...)
	router.PUT(options.BaseURL+"/trainer/calendar/make-hour-unavailable", wrapper.MakeHourUnavailable, options.OperationMiddlewares["makeHourUnavailable"]...)

}

type GetTrainerAvailableHoursRequestObject struct {
	Params GetTrainerAvailableHoursParams
}

type GetTrainerAvailableHoursResponseObject interface {
	VisitGetTrainerAvailableHoursResponse(w http.ResponseWriter) error
}

type GetTrainerAvailableHours200JSONResponse []Date

func (response GetTrainerAvailableHours200JSONResponse) VisitGetTrainerAvailableHoursResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetTrainerAvailableHoursdefaultJSONResponse struct {
	Body       ErrorResponse
	StatusCode int
}

func (response GetTrainerAvailableHoursdefaultJSONResponse) VisitGetTrainerAvailableHoursResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type MakeHourAvailableRequestObject struct {
	Body *MakeHourAvailableJSONRequestBody
}

type MakeHourAvailableResponseObject interface {
	VisitMakeHourAvailableResponse(w http.ResponseWriter) error
}

type MakeHourAvailable204JSONResponse []Date

func (response MakeHourAvailable204JSONResponse) VisitMakeHourAvailableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(204)
	_, err := buf.WriteTo(w)
	return err
}

type MakeHourAvailabledefaultJSONResponse struct {
	Body       ErrorResponse
	StatusCode int
}

func (response MakeHourAvailabledefaultJSONResponse) VisitMakeHourAvailableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type MakeHourUnavailableRequestObject struct {
	Body *MakeHourUnavailableJSONRequestBody
}

type MakeHourUnavailableResponseObject interface {
	VisitMakeHourUnavailableResponse(w http.ResponseWriter) error
}

type MakeHourUnavailable204JSONResponse []Date

func (response MakeHourUnavailable204JSONResponse) VisitMakeHourUnavailableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(204)
	_, err := buf.WriteTo(w)
	return err
}

type MakeHourUnavailabledefaultJSONResponse struct {
	Body       ErrorResponse
	StatusCode int
}

func (response MakeHourUnavailabledefaultJSONResponse) VisitMakeHourUnavailableResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {

	// (GET /trainer/calendar)
	GetTrainerAvailableHours(ctx context.Context, request GetTrainerAvailableHoursRequestObject) (GetTrainerAvailableHoursResponseObject, error)

	// (PUT /trainer/calendar/make-hour-available)
	MakeHourAvailable(ctx context.Context, request MakeHourAvailableRequestObject) (MakeHourAvailableResponseObject, error)

	// (PUT /trainer/calendar/make-hour-unavailable)
	MakeHourUnavailable(ctx context.Context, request MakeHourUnavailableRequestObject) (MakeHourUnavailableResponseObject, error)
}

type StrictHandlerFunc func(ctx echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetTrainerAvailableHours operation middleware
func (sh *strictHandler) GetTrainerAvailableHours(ctx echo.Context, params GetTrainerAvailableHoursParams) error {
	var request GetTrainerAvailableHoursRequestObject

	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetTrainerAvailableHours(ctx.Request().Context(), request.(GetTrainerAvailableHoursRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetTrainerAvailableHours")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetTrainerAvailableHoursResponseObject); ok {
		return validResponse.VisitGetTrainerAvailableHoursResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// MakeHourAvailable operation middleware
func (sh *strictHandler) MakeHourAvailable(ctx echo.Context) error {
	var request MakeHourAvailableRequestObject

	var body MakeHourAvailableJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.MakeHourAvailable(ctx.Request().Context(), request.(MakeHourAvailableRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "MakeHourAvailable")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(MakeHourAvailableResponseObject); ok {
		return validResponse.VisitMakeHourAvailableResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// MakeHourUnavailable operation middleware
func (sh *strictHandler) MakeHourUnavailable(ctx echo.Context) error {
	var request MakeHourUnavailableRequestObject

	var body MakeHourUnavailableJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.MakeHourUnavailable(ctx.Request().Context(), request.(MakeHourUnavailableRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "MakeHourUnavailable")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(MakeHourUnavailableResponseObject); ok {
		return validResponse.VisitMakeHourUnavailableResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}